Update README.md
# Project Title 🚀   # tribyut-page

## Project Description 📝

> This is a project which gives information about Benazir Buto.

Examples:

It is just a class project.

```html
<main id = "main">
    <link rel="stylesheet" href="styles.css">
<h1 id="title">Benazir Botu</h1>
<figure id = "img-div">
    <img src = "https://media.gettyimages.com/id/50432776/photo/benazir-bhutto.jpg?s=612x612&w=gi&k=20&c=tGiJIj9lF_DMBSdzWfTs8vC6ppQ3frokDSENvqTh-mA=" id="image">
<div id="img-caption">
    Benazir Bhutto, the Prime Minister of Pakistan.
</div>
<p id="tribute-info">
Benazir Bhutto[a] (21 June 1953 – 27 December 2007) was a Pakistani politician and stateswoman who served as the 11th and 13th prime minister of Pakistan from 1988 to 1990 and again from 1993 to 1996. She was the first woman elected to head a democratic government in a Muslim-majority country. Ideologically a liberal and a secularist, she chaired or co-chaired the Pakistan People's Party (PPP) from the early 1980s until her assassination in 2007.
</p>
<a href="https://en.wikipedia.org/wiki/Benazir_Bhutto" rel="noopener" id="tribute-link" target ="_blank">Wikipedia</a>
</figure>
</main>
```

```css
#image {
  max-width: 100%;
  height: auto;
  margin: 0 auto;
  display: block;
}

```


## Demo 📸

file:///C:/Users/Rezaie/Desktop/tribute_page/tribyut-page/index.html

## Technologies Used 🛠️

- HTML
- CSS

## Author 👩‍💻
Sharifa Rezai
- Email: [example](sharifashaida82@gmail.com)